ToyExecutionEnvironmentV2, testing module for common code
