+++ arithmetization/build.gradle
@@ -34,7 +34,6 @@ apply from: rootProject.file("gradle/lint.gradle")
 apply from: rootProject.file("gradle/trace-files.gradle")
 
 dependencies {
-  testImplementation project(path: ':reference-tests')
   // annotationProcessor generates the file META-INF/services/org.hyperledger.besu.plugin.BesuPlugin
   annotationProcessor 'com.google.auto.service:auto-service'
   compileOnly "${besuArtifactGroup}:evm"
@@ -61,6 +60,8 @@ dependencies {
   implementation 'io.tmio:tuweni-toml'
   implementation 'org.bouncycastle:bcprov-jdk18on'
   implementation 'org.hibernate.validator:hibernate-validator'
+
+  testImplementation project(path: ':testing')
 
   testImplementation "${besuArtifactGroup}:evm"
   testImplementation "${besuArtifactGroup}:besu-datatypes"
