+++ gradle.properties
@@ -1,5 +1,5 @@
 releaseVersion=0.3.0-rc1
-besuVersion=24.7-develop-f812936
+besuVersion=24.7-develop-5622666
 besuArtifactGroup=io.consensys.linea-besu
 distributionIdentifier=linea-arithmetization
 distributionBaseUrl=https://artifacts.consensys.net/public/linea-besu/raw/names/linea-besu.tar.gz/versions/
