+++ testing/src/main/java/net/consensys/linea/GeneralStateTestCaseEipSpec.java
@@ -13,8 +13,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.zktracer.testing;
+package net.consensys.linea;
 
+import java.util.List;
 import java.util.function.Supplier;
 
 import org.hyperledger.besu.datatypes.Hash;
@@ -35,7 +36,7 @@ public class GeneralStateTestCaseEipSpec {
   // anything
   // is run, which isn't friendly and 2) this makes it harder to parallelize this step. Anyway, this
   // is why this is a supplier: calling get() actually does the signing.
-  private final Supplier<Transaction> transactionSupplier;
+  private final List<Supplier<Transaction>> transactionSuppliers;
 
   private final ReferenceTestWorldState initialWorldState;
 
@@ -51,9 +52,9 @@ public class GeneralStateTestCaseEipSpec {
   private final int valueIndex;
   private final String expectException;
 
-  GeneralStateTestCaseEipSpec(
+  public GeneralStateTestCaseEipSpec(
       final String fork,
-      final Supplier<Transaction> transactionSupplier,
+      final List<Supplier<Transaction>> transactionSuppliers,
       final ReferenceTestWorldState initialWorldState,
       final Hash expectedRootHash,
       final Hash expectedLogsHash,
@@ -63,7 +64,7 @@ public class GeneralStateTestCaseEipSpec {
       final int valueIndex,
       final String expectException) {
     this.fork = fork;
-    this.transactionSupplier = transactionSupplier;
+    this.transactionSuppliers = transactionSuppliers;
     this.initialWorldState = initialWorldState;
     this.expectedRootHash = expectedRootHash;
     this.expectedLogsHash = expectedLogsHash;
@@ -90,9 +91,13 @@ public class GeneralStateTestCaseEipSpec {
     return expectedLogsHash;
   }
 
-  public Transaction getTransaction() {
+  public int getTransactionsCount() {
+    return transactionSuppliers.size();
+  }
+
+  public Transaction getTransaction(int txIndex) {
     try {
-      return transactionSupplier.get();
+      return transactionSuppliers.get(txIndex).get();
     } catch (RuntimeException re) {
       // some tests specify invalid transactions.  We throw exceptions in
       // GeneralStateTests but they are encoded in BlockchainTests, so we
